modified the content in 05 and 06 folder

+++ 06 - PyTorch基本语法/C01Tensors.py
@@ -1,10 +1,8 @@
 import torch
-import torch.nn as nn
-import torch.optim as optim
 
 
-alist = [1.0, 2, 3, 4, 5]
-tensor = torch.tensor(alist)
+alist = [1, 2, 3, 4, 5]
+tensor = torch.tensor(alist, dtype=torch.float)
 print(tensor)
 
 zeros_tensor = torch.zeros(5)
@@ -16,8 +14,6 @@ print(zeros_tensor)
 print(ones_tensor)
 print(random_tensor)
 print(normal_tensor)
-
-# print(zeros_tensor.item())  # ValueError: only one element tensors can be converted to Python scalars
 
 sliced_tensor = normal_tensor[:3]
 print(sliced_tensor)
